+++ jdplus-main-desktop/jdplus-toolkit-desktop-plugin/src/main/java/jdplus/toolkit/desktop/plugin/ui/variables/VariablesDocumentManager.java
@@ -124,6 +124,10 @@ public class VariablesDocumentManager extends AbstractWorkspaceItemManager<TsDat
         return WorkspaceItem.system(ID, name, p);
     }
 
+    public static WorkspaceItem<TsDataSuppliers> newItem(String name, TsDataSuppliers p) {
+        return WorkspaceItem.newItem(ID, name, p);
+    }
+
     @Override
     public boolean isAutoLoad() {
         return true;