+++ jdplus-main-desktop/jdplus-toolkit-desktop-plugin/src/main/java/jdplus/toolkit/desktop/plugin/ui/variables/actions/RemoveVariablesAction.java
@@ -61,7 +61,7 @@ public final class RemoveVariablesAction extends SingleNodeAction<ItemWsNode> {
     }
 
     @Messages({
-        "RemoveVariables.dialog.title=Remove calendar",
+        "RemoveVariables.dialog.title=Remove variables",
         "RemoveVariables.dialog.message=Are you sure?"
     })
     static void removeVariables(TsDataSuppliers p, ItemWsNode node) {